refactor: 회원 탈퇴 시 스케쥴러에서 처리하지 않도록 수정
